Tianocore [DEBUG] Remove off-by-one in SnpDxe driver initialization
I am not sure if this is supposed to be off-by-one or not, but it seems very
suspicious. The Tianocore code upstream implies this might be "just fine", but
it's very odd if so. So I am making this quick one-liner while I figure out
the Dxe code. I have no idea really about this part of Tianocore at all yet.

Signed-off-by: Jon Masters <jcm@redhat.com>