Fix handling of BootNext in BDS
Create a default boot order if one does not exist. Make sure the
console is enabled before any banner is printed.

Currently, code exists to delete the BootNext image, but it is supposed
to be called after the image is booted and is never actually called.
Move that code to before image boot so that BootNext is properly deleted.

Delete the BootNext variable after it is read but before actually booting
the image.

Signed-off-by: Mark Salter <msalter@redhat.com>