Fix Arch64 booting with boot entries created by efibootmgr
The efibootmgr boot entries contain a partial device path to
the boot file beginning with partition. Entries created via
Boot Manager would start with an ACPI path followed by PCI
and SATA paths. The ARM BDS code would fill in the missing
bits as if the partition were on a removeable device. This
allowed UEFI to boot the application (grub). The problem
was that the original shortened device path was passed to the
LoadImage interface and LoadImage was unable to find a
DeviceHandle for it. Grub would then not be able to find
the disk where its config file was located which caused the
boot to stop and grub to enter its shell. This patch fixes
this by passing the full path along to LoadImage so that a
proper DeviceHandle is passed to the application using the
Loaded Image Protocol.

Signed-off-by: Mark Salter <msalter@redhat.com>