Add manual dtb file build scripts and infrastructure
Add the APM Mustang dts files to the repository under a DeviceTree
directory, and add a script that needs to be manually run to convert the
dts files in to a DTB encoded as a UINT8 array called DeviceTreeBuff for
consumption by other parts of UEFI.

Also add the DeviceTree.inf file. The referenced DeviceTree.c file is
built by the script and isn't under revision control for now.

Some future patch may add support for calling makedtb.sh as part of
the `make tianocore` build process.

Signed-off-by: Mark Langsdorf <mlangsdo@redhat.com>